python/requirements.txt
pytesseract
pyzenhub
